requirements.txt
Pillow
audiotsm
scipy
numpy
pytube3
